Split note/sender BT nodes into subpackages
