+++ vultron/core/behaviors/note/nodes/__init__.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""BT nodes subpackage for note workflows.
+
+Re-exports node classes from semantic submodules so existing import paths
+(``from vultron.core.behaviors.note.nodes import ...``) continue to work.
+"""
+
+from vultron.core.behaviors.note.nodes.creation import (
+    AttachNoteFromResultNode,
+    CreateNoteNode,
+)
+from vultron.core.behaviors.note.nodes.storage import (
+    AttachNoteToCaseNode,
+    SaveNoteNode,
+)
+
+__all__ = [
+    "SaveNoteNode",
+    "AttachNoteToCaseNode",
+    "CreateNoteNode",
+    "AttachNoteFromResultNode",
+]
